♻️ refactor: type 키워드에서 interface 키워드로 수정
-  다른 파일에서 객체 타입 선언할 때 interface를 사용해서, 코드를 일관적으로 작성하기 위해 type 키워드를 interface 키워드로 수정합니다
