+++ components/NotificationModal/NotificationModal.tsx
@@ -4,7 +4,7 @@ import closeIcon from "@/public/images/close_icon.svg";
 import styled from "@emotion/styled";
 import Image from "next/image";
 import Notification from "./Notification";
-import { NotificationItem } from "./types/type";
+import { NotificationItem } from "./types/types";
 
 export default function NotificationModal({
   isModalVisible,
